+++ lib/EasyPost/AddressVerificationFieldError.php
@@ -4,10 +4,11 @@ namespace EasyPost;
 
 /**
  * @package EasyPost
- * @property string $field
  * @property string $message
+ * @property string $code
+ * @property string $field
  * @property string $suggestion
  */
-class FieldError extends EasyPostObject
+class AddressVerificationFieldError extends EasyPostObject
 {
 }