+++ lib/EasyPost/VerificationDetails.php
@@ -4,9 +4,9 @@ namespace EasyPost;
 
 /**
  * @package EasyPost
- * @param int $latitude
- * @param int $longitude
- * @param string $time_zone
+ * @property int $latitude
+ * @property int $longitude
+ * @property string $time_zone
  */
 class VerificationDetails extends EasyPostObject
 {