+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Adds `custom_headers` property to webhook model
 - Corrects wrapping payload for update webhook endpoint
+- Fixes error handling
+  - Corrects the type of `errors` property on an `ApiException` to allow for the alternative format
+  - Fixes a bug where you could not `prettyPrint` an error if it used the alternative format
+  - Corrects the `param` references to `property` references on all error models
+  - Adds new `AddressVerificationFieldError` for the `errors` property on a `Verification` model
+  - Adds missing `suggestion` property to `FieldError`
 - Removes the deprecated `create_list` tracker endpoint function as it is no longer available via API
 
 ## v7.4.2 (2024-08-16)